feat(meetings): add reusable open_meeting_picker command
